refactor: moved session provider to auth folder
